refactor: move `storage` to `pkg/`
