+++ internal/server/server.go
@@ -7,11 +7,11 @@ import (
 
 	"github.com/PitiNarak/condormhub-backend/internal/database"
 	"github.com/PitiNarak/condormhub-backend/internal/middleware"
-	"github.com/PitiNarak/condormhub-backend/internal/storage"
 	"github.com/PitiNarak/condormhub-backend/pkg/apperror"
 	"github.com/PitiNarak/condormhub-backend/pkg/email"
 	"github.com/PitiNarak/condormhub-backend/pkg/jwt"
 	"github.com/PitiNarak/condormhub-backend/pkg/redis"
+	"github.com/PitiNarak/condormhub-backend/pkg/storage"
 	"github.com/PitiNarak/condormhub-backend/pkg/stripe"
 	"github.com/goccy/go-json"
 	"github.com/gofiber/fiber/v2"
